BlogComment/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
configparser==3.8.1
coverage==6.2
Django==3.2.12
django-tinymce==3.4.0
django-widget-tweaks==1.4.11
EasyProcess==1.1
flake8==4.0.1
future==0.18.2
idna==2.10
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.6.1
MeaningCloud-python==2.0.0
Pillow==6.2.2
pkg-resources==0.0.0
protobuf==3.19.4
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.8
zipp==3.6.0
